finish cache service & start teams

+++ services/user/routes/delete_api_key.go
@@ -1,12 +1,15 @@
 package routes
 
 import (
-	"net/http"
-	"io"
 	"crypto/sha256"
 	"encoding/base64"
-	"errors"
 	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"strings"
 
 	"gorm.io/gorm"
 )
@@ -15,6 +18,10 @@ type DeleteApiKeyRequest struct {
 	Name string `json:"name"`
 	Username string `json:"username"`
 	Password string `json:"password"`
+}
+
+type InvalidateRequest struct {
+	Hash string `json:"hash"`
 }
 
 func DeleteApiKey(w http.ResponseWriter, r *http.Request, db gorm.DB) error {
@@ -60,20 +67,29 @@ func DeleteApiKey(w http.ResponseWriter, r *http.Request, db gorm.DB) error {
 	apiKey := ApiKey{}
 
 	for i, v := range iotUser.ApiKeys {
-		key := ApiKey{}
+		if v == deleteApiKeyRequest.Name {
+			key := ApiKey{}
 
-		err = db.First(&key, v).Error
+			err = db.Where("name = ?", v).First(&key).Error
 
-		if err != nil {
-			return err
+			if err != nil {
+				return err
+			}
+
+			if apiKey.Name != "" {
+				index = i
+			}
+			
+			apiKey = key
 		}
-
-		if apiKey.Name != "" {
-			index = i
-		}
-		
-		apiKey = key
 	} 
+
+	if apiKey.Name == "" {
+		w.WriteHeader(404)
+		w.Write([]byte("This API key does not exist!"))
+
+		return nil
+	}
 
 	err = db.Delete(&apiKey).Error
 
@@ -85,6 +101,43 @@ func DeleteApiKey(w http.ResponseWriter, r *http.Request, db gorm.DB) error {
 		iotUser.ApiKeys[:index],
 		iotUser.ApiKeys[index + 1:]...,
 	)
+
+	fmt.Println(iotUser)
+
+	invalidateRequest := InvalidateRequest{
+		Hash: encoded,
+	}	
+
+	stringified, err := json.Marshal(invalidateRequest)
+
+	if err != nil {
+		return err
+	}
+
+	host := os.Getenv("IOT_CACHE_SERVICE_HOST")	
+	port := os.Getenv("IOT_CACHE_SERVICE_PORT_PROXY")
+
+	req, err := http.NewRequest(
+		"POST",
+		fmt.Sprintf("http://%s:%s/invalidate", host, port),
+		strings.NewReader(string(stringified)),	
+	)
+
+	if err != nil {
+		return err
+	}
+
+	res, err := http.DefaultClient.Do(req)
+
+	if err != nil {
+		return err
+	}
+
+	if res.StatusCode != 200 {
+		http.Error(w, "There was an error deleting the API key.", http.StatusInternalServerError)
+
+		return nil
+	}
 
 	err = db.Save(&iotUser).Error
